+++ src/onegov.activity/onegov/activity/tests/test_iso20022.py
@@ -1,0 +1,87 @@
+from datetime import date
+from decimal import Decimal
+from onegov.activity.iso20022 import extract_transactions
+
+
+def test_extract_transactions(postfinance_xml):
+    transactions = extract_transactions(postfinance_xml)
+
+    t = next(transactions)
+    assert t.booking_date == date(2016, 4, 30)
+    assert t.valuta_date == date(2016, 4, 30)
+    assert t.amount == Decimal('24.00')
+    assert t.currency == 'CHF'
+    assert t.debitor is None
+    assert t.debitor_account is None
+    assert t.credit is False
+    assert t.reference is None
+    assert t.booking_text \
+        == "UEBRIGE: 25-9034-2 FÜR KONTOAUSZUG PAPIER TÄGLICH"
+
+    t = next(transactions)
+    assert t.amount == Decimal('328.75')
+
+    t = next(transactions)
+    assert t.amount == Decimal('200.80')
+
+    t = next(transactions)
+    assert t.amount == Decimal('638.15')
+
+    t = next(transactions)
+    assert t.amount == Decimal('24.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('35.72')
+
+    t = next(transactions)
+    assert t.amount == Decimal('100.00')
+    assert t.reference == '20160526531801000100274'
+
+    t = next(transactions)
+    assert t.amount == Decimal('50.00')
+    assert t.debitor == "Rutschmann Pia"
+    assert t.debitor_account is None
+
+    t = next(transactions)
+    assert t.amount == Decimal('200.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('50.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('100.00')
+    assert t.credit is True
+
+    t = next(transactions)
+    assert t.amount == Decimal('200.00')
+    assert t.debitor == "Bernasconi Maria"
+    assert t.debitor_account == "CH4444444444444444444"
+
+    t = next(transactions)
+    assert t.amount == Decimal('50.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('100.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('50.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('100.00')
+    assert t.reference == "160527CH00T99EBW"
+
+    t = next(transactions)
+    assert t.amount == Decimal('50.00')
+    assert t.note == "MUSTER MITTEILUNG"
+
+    t = next(transactions)
+    assert t.amount == Decimal('20.00')
+
+    t = next(transactions)
+    assert t.amount == Decimal('100.00')
+    assert t.note == '\n'.join([
+        "MUSTER MITTEILUNG 1",
+        "MUSTER MITTEILUNG 2",
+        "MUSTER MITTEILUNG 3",
+        "MUSTER MITTEILUNG 4",
+    ])